production/requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==6.0.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
